+++ modules/OpticFlow/land_guidance.h
@@ -39,8 +39,8 @@ void run_land_guidance_onvision(void);
 extern int32_t vision_pgain;
 extern int32_t vision_dgain;
 
-extern int32_t vision_land_pgain;
-extern int32_t vision_land_dgain;
+extern float vision_land_pgain;
+extern float vision_land_dgain;
 extern float vision_div_const;
 
 #endif /* LAND_GUIDANCE_H_ */
